subdivide tests between forward and backward
