+++ Component/ODESolver/Forward/tests/CentralDifferenceExplicitSolverDynamic_test.cpp
@@ -24,9 +24,11 @@ using sofa::testing::BaseSimulationTest;
 #include <sofa/testing/NumericTest.h>
 using sofa::testing::NumericTest;
 
-#include <SofaExplicitOdeSolver_test/MassSpringSystemCreation.h>
+#include <sofa/component/odesolver/testing/MassSpringSystemCreation.h>
 
 #include <SceneCreator/SceneCreator.h>
+
+
 
 //Including Simulation
 #include <sofa/simulation/Simulation.h>
@@ -38,10 +40,7 @@ using sofa::testing::NumericTest;
 using MechanicalObject3 = sofa::component::container::MechanicalObject<sofa::defaulttype::Vec3Types> ;
 
 // Solvers
-#include <SofaGeneralExplicitOdeSolver/RungeKutta4Solver.h>
-#include <SofaBaseLinearSolver/CGLinearSolver.h>
-
-#include <sofa/defaulttype/VecTypes.h>
+#include <SofaGeneralExplicitOdeSolver/CentralDifferenceSolver.h>
 
 namespace sofa {
 
@@ -56,27 +55,18 @@ Test the dynamic behavior of solver: study a mass-spring system under gravity in
 The movement follows the equation:
 x(t)= A cos(wt + phi) with w the pulsation w=sqrt(K/M), K the stiffness, M the mass and phi the phase.
 In this test: x(t=0)= 1 and v(t=0)=0 and K = spring stiffness and phi = 0 of material thus x(t)= cos(wt)
-This tests generates the discrete mass position obtained with rungeKutta4 solver with different parameter values (K,M,h).
+This tests generates the discrete mass position obtained with central difference solver with different parameter values (K,M,h).
 Then it compares the effective mass position to the computed mass position every time step.
 */
 
 template <typename _DataTypes>
-struct RungeKutta4ExplicitSolverDynamic_test : public BaseSimulationTest
+struct CentralDifferenceExplicitSolverDynamic_test : public BaseSimulationTest
 {
-
-
     typedef _DataTypes DataTypes;
-    typedef typename DataTypes::CPos CPos;
     typedef typename DataTypes::Coord Coord;
-    typedef typename DataTypes::VecCoord VecCoord;
-    typedef typename DataTypes::VecDeriv VecDeriv;
-    typedef typename DataTypes::Deriv Deriv;
-    typedef typename DataTypes::Real Real;
 
     typedef container::MechanicalObject<DataTypes> MechanicalObject;
-    typedef component::odesolver::RungeKutta4Solver RungeKutta4Solver;
-    typedef component::linearsolver::CGLinearSolver<component::linearsolver::GraphScatteredMatrix, component::linearsolver::GraphScatteredVector> CGLinearSolver;
-
+    typedef component::odesolver::CentralDifferenceSolver CentralDifferenceSolver;
 
     /// Root of the scene graph
     simulation::Node::SPtr root;      
@@ -88,7 +78,7 @@ struct RungeKutta4ExplicitSolverDynamic_test : public BaseSimulationTest
     vector<double> accelerationsArray;
     
     /// Create the context for the scene
-    void createScene(double K, double m, double l0)
+    void createScene(double K, double m, double l0,double rm)
     {
         // Init simulation
         sofa::simulation::setSimulation(simulation = new sofa::simulation::graph::DAGSimulation());
@@ -98,12 +88,8 @@ struct RungeKutta4ExplicitSolverDynamic_test : public BaseSimulationTest
         root->setGravity(Coord(0,-10,0));
 
         // Solver
-        RungeKutta4Solver::SPtr rungeKutta4Solver = addNew<RungeKutta4Solver> (root);
-
-        CGLinearSolver::SPtr cgLinearSolver = addNew<CGLinearSolver> (root);
-        cgLinearSolver->d_maxIter.setValue(3000);
-        cgLinearSolver->d_tolerance.setValue(1e-9);
-        cgLinearSolver->d_smallDenominatorThreshold.setValue(1e-9);
+        CentralDifferenceSolver::SPtr centralDifferenceSolver = addNew<CentralDifferenceSolver> (root);
+        centralDifferenceSolver->f_rayleighMass.setValue(double(rm));
 
         // Set initial positions and velocities of fixed point and mass
         MechanicalObject3::VecCoord xFixed(1);
@@ -129,7 +115,7 @@ struct RungeKutta4ExplicitSolverDynamic_test : public BaseSimulationTest
 
 
     /// Generate discrete mass position values with euler implicit solver
-    void generateDiscreteMassPositions (double h, double K, double m, double z0, double v0,double g, double finalTime)
+    void generateDiscreteMassPositions (double h, double K, double m, double z0, double v0,double g, double finalTime, double rm)
     {
         int size = 0 ;
 
@@ -151,49 +137,42 @@ struct RungeKutta4ExplicitSolverDynamic_test : public BaseSimulationTest
         // First velocity is v0
         velocitiesArray.push_back(v0);
 
-        // Constants
-        double stepBy2 = h/2.0;
-        double stepBy3 = h/3.0;
-        double stepBy6 = h/6.0;
-        double pos,vel,newX,k1v,k2v,k3v,k4v,k1a,k2a,k3a,k4a,pos2,vel2,acc2;
+        double pos,acc,vel;
 
-        for(int i=1;i< size+1; i++)
+        if(rm == 0)
         {
+            for(int i=1;i< size+1; i++)
+            {
+                acc = (-K*(positionsArray[i-1]-z0) - m*g)/m;
+                vel = velocitiesArray[i-1]+acc*h;
+                pos = positionsArray[i-1]+vel*h;
 
-            // Step 1
-            pos = positionsArray[i-1];
-            vel = velocitiesArray[i-1];
-            k1v = vel;
-            k1a = accelerationsArray[i-1];
+                // Fill array
+                positionsArray.push_back(pos);
+                accelerationsArray.push_back(acc);
+                velocitiesArray.push_back(vel);
 
-            // Step 2
-            newX = pos + k1v*stepBy2;
-            k2v  = vel + k1a*stepBy2;
-            k2a  = (-K*(newX-z0)-m*g)/m;
+            }
 
-            // Step 3
-            newX = pos + k2v * stepBy2;
-            k3v  = vel + k2a * stepBy2;
-            k3a  = (-K*(newX-z0)-m*g)/m;
+        }
 
-            // Step 4
-            newX = pos + k3v * h;
-            k4v  = vel + k3a * h;
-            k4a  = (-K*(newX-z0)-m*g)/m;
+        else
+        {
+            double constantVel =  ((1.0/h) - (rm/2.0))/((1.0/h) + (rm/2.0));
+            double constantAcc = 1.0/((1.0/h) + (rm/2.0));
 
-            // Position pos2 at time t+h
-            pos2 = pos + k1v * stepBy6 + k2v * stepBy3 + k3v * stepBy3 + k4v * stepBy6;
+            for(int i=1;i< size+1; i++)
+            {
+               acc = (-K*(positionsArray[i-1]-z0) - m*g)/m;
+               vel = velocitiesArray[i-1] * constantVel  + acc*constantAcc;
+               pos = positionsArray[i-1]+vel*h;
 
-            // Velocity vel2 at time t+h
-            vel2 = vel + k1a * stepBy6 + k2a * stepBy3 + k3a * stepBy3 + k4a * stepBy6;
+                // Fill array
+                positionsArray.push_back(pos);
+                accelerationsArray.push_back(acc);
+                velocitiesArray.push_back(vel);
 
-            // Acceleration at time t+h
-            acc2 = (-K*(pos2-z0)-m*g)/m;
-
-            // Fill array
-            positionsArray.push_back(pos2);
-            velocitiesArray.push_back(vel2);
-            accelerationsArray.push_back(acc2);
+            }
 
         }
 
@@ -248,30 +227,30 @@ typedef Types<
 > DataTypes; // the types to instanciate.
 
 // Test suite for all the instanciations
-TYPED_TEST_SUITE(RungeKutta4ExplicitSolverDynamic_test, DataTypes);
+TYPED_TEST_SUITE(CentralDifferenceExplicitSolverDynamic_test, DataTypes);
 
-// Test case: h=0.1 k=100 m =10 rm=0.1 rk=0.1
-TYPED_TEST( RungeKutta4ExplicitSolverDynamic_test , rungeKutta4ExplicitSolverDynamicTest_medium_dt)
+// Test case: h=0.01 k=100 m =10 rm=0.1 rk=0.1
+TYPED_TEST( CentralDifferenceExplicitSolverDynamic_test , centralDifferenceExplicitSolverDynamicTest_medium_dt_without_damping)
 {
-   this->createScene(100,10,1); // k,m,l0
-   this->generateDiscreteMassPositions (0.01, 100, 10, 1, 0, 10, 2);
-   this-> compareSimulatedToTheoreticalPositions(5e-16,0.01);
+   this->createScene(100,10,1,0); // k,m,l0
+   this->generateDiscreteMassPositions (0.01, 100, 10, 1, 0, 10, 2, 0);
+   this-> compareSimulatedToTheoreticalPositions(4e-16,0.01);
 }
 
 // Test case: h=0.1 K=1000 m = 10
-TYPED_TEST( RungeKutta4ExplicitSolverDynamic_test , rungeKutta4ExplicitSolverDynamicTest_high_dt)
+TYPED_TEST( CentralDifferenceExplicitSolverDynamic_test , rungeKutta2ExplicitSolverDynamicTest_high_dt_with_damping)
 {
-   this->createScene(1000,10,1); // k,m,l0
-   this->generateDiscreteMassPositions (0.1, 1000, 10, 1, 0, 10, 2);
+   this->createScene(1000,10,1,0.1); // k,m,l0
+   this->generateDiscreteMassPositions (0.1, 1000, 10, 1, 0, 10, 2, 0.1);
    this-> compareSimulatedToTheoreticalPositions(5e-16,0.1);
 }
 
 // Test case: h=0.001 k=100 m=100
-TYPED_TEST( RungeKutta4ExplicitSolverDynamic_test , rungeKutta4ExplicitSolverDynamicTest_small_dt)
+TYPED_TEST( CentralDifferenceExplicitSolverDynamic_test , rungeKutta2ExplicitSolverDynamicTest_small_dt_with_damping)
 {
-   this->createScene(100,100,1); // k,m,l0
-   this->generateDiscreteMassPositions (0.001, 100, 100, 1, 0, 10, 2);
-   this-> compareSimulatedToTheoreticalPositions(9e-16,0.001);
+   this->createScene(100,100,1,0.1); // k,m,l0
+   this->generateDiscreteMassPositions (0.001, 100, 100, 1, 0, 10, 2, 0.1);
+   this-> compareSimulatedToTheoreticalPositions(2e-16,0.001);
 }
 
 } // namespace sofa
